Rename 'New' to 'Make'
SpTOL is a performance critical library. It is better that we can minimize
malloc's. Therefore, making a sptVector can be done directly on the stack. Also
it makes easier to embed sptVector into other data structures.
By renaming 'New' to 'Make', it tells the caller that we are not allocating the
sptVector itself on the heap, similar to having an std::vector on the stack.
The name 'make' comes from Go, where it means to produce a complex data
structure.